requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
Pillow==7.2.0
python-pptx==0.6.18
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.9
XlsxWriter==1.2.9
